chore: 버튼 pointer 및 세밀한 css 수정
