--- src/pages/Apply/JobNoticeApply.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import JobNoticeApplyCenter from './JobNoticeApplyCenter';
-import Header from '../../components/Header';
-const JobNoticeApply = () => {
-  return (
-    <div>
-      <JobNoticeApplyCenter />
-    </div>
-  );
-};
-
-export default JobNoticeApply;
